Updated Regulus.py
