+++ regulus/math/inv_kernel_reg.py
